OMG, tests not all passing. Bad developer!
* However, Dispatcher, MagicMethods, and many delegate types are now entirely autogenerated. Most changes are direct consequences of this.
** Missing tests are marked failing, in retrievetypetest; various deleted tests are now redundant or meaningless.
** generatedispatcher could probably use some sort of direct testing...
* As it turned out, I didn't need P25M.NullObject. Bah.
* Numpy behaviour has changed slightly, but I'm not sure which version is correct. Definitely worth checking in at this point for further tests and investigation elsewhere.

by: William

+++ src/Python25Mapper_tuple.cs
@@ -67,7 +67,7 @@ namespace Ironclad
         public override int
         PyTuple_Size(IntPtr tuplePtr)
         {
-            return CPyMarshal.ReadIntField(tuplePtr, typeof(PyTupleObject), "ob_size");
+            return (int)CPyMarshal.ReadIntField(tuplePtr, typeof(PyTupleObject), "ob_size");
         }
         
         public override IntPtr
